api/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "A simple FastAPI application"
authors = [{ name = "444chak", email = "444chak@gmail.com" },
            { name = "Eliott B", email = "contact@eliott-b.fr"}]
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.11"
readme = "../README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
api = "uvicorn src.app:app --reload --log-level debug"

[tool.ruff]
select = ["ALL"]